+++ tests/test_protocol_encode_dummy.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
-#  strftime.py
+#  test_protocol_decode_json.py
 #
 #  Copyright 2017 Jelle Smet <development@smetj.net>
 #
@@ -22,25 +22,11 @@
 #
 #
 
-from wishbone.function.template import TemplateFunction
-import arrow
+
+from wishbone.protocol.encode.dummy import Dummy
 
 
-class STRFTime(TemplateFunction):
+def test_protocol_encode_dummy():
 
-    '''
-    **Returns a formatted version of an epoch timestamp.**
-
-    - Parameters to initialize the function:
-
-        None
-
-    - Parameters to call the function:
-
-        - epoch(int/float)(): Epoch value
-        - template(str)(): The template to which to convert the provided epoch
-    '''
-
-    def lookup(self, epoch, template):
-
-        return arrow.get(epoch).format(template)
+    m = Dummy()
+    assert m.handler({"one": 1}) == {"one": 1}